+++ mocks/fileMock.js
@@ -1,0 +1,1 @@
+module.exports = 'test-file-stub';
