+++ server/index.js
@@ -5,7 +5,7 @@ const compression = require('compression');
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: true}));
 app.use(compression());
-const Reviews = require('../database/Reviews.js');
+const Reviews = require('../database/Reviews');
 app.set('PORT', 1337);
 app.use(express.static(__dirname + '/../client/dist'));
