+++ webpack.common.js
@@ -53,16 +53,4 @@ module.exports = {
       chunkFilename: "stylesAsync.css"
     })
   ],
-  // module: {
-  //   rules: [
-  //     {
-  //       test: /\.css$/,
-  //       use: [
-  //         MiniCssExtractPlugin.loader,
-  //         "css-loader"
-  //       ]
-  //     }
-  //   ]
-  // }
-
 };
